Version 1.3.0 Feat: Toggle Status feature & misc
Added addEventListener on readButtonElem which loops through the
mainArr node list. Once nodelist is equal to the specific card, two 'if
statements' trigger if the obj value of the library[i].status is equal
to 'true' or 'false'. It will change the value to the respective
boolean value if the user has read the book or not.

misc: added img-icons on assets file.